Add config for the gorestful middleware
One of the benefits of using gorestful is its fancy routes, which
let you specify "path parameters" such as "/users/{id}".  This
change adds a "Config" structure for the gorestful middleware that
allows you to use the gorestful route path and not the actual
request path.

For backward compatibility, this introduces the "HandlerWithConfig"
function, and the existing "Handler" function calls "HandlerWithConfig"
with the default "Config".

The example has been updated to use "HandlerWithConfig".